conflict

+++ src/main/java/com/playdata/pdfolio/domain/response/gather/GatherDetailResponse.java
@@ -25,8 +25,8 @@
         private Long teamSize;
         private GatherCategory category;
         private String contact;
-        private Long heartCount;
-        private Long viewCount;
+        private Integer heartCount;
+        private Integer viewCount;
         private Boolean isDeleted;
 
         private Member member;
