+++ src/main/java/com/playdata/pdfolio/domain/response/gather/GatherResponse.java
@@ -27,8 +27,8 @@ public class GatherResponse {
     private Long teamSize;
     private GatherCategory category;
     private String contact;
-    private Long heartCount;
-    private Long viewCount;
+    private Integer heartCount;
+    private Integer viewCount;
     private Boolean isDeleted;
 
     private Member member;
